drf_api/urls.py
from django.contrib import admin
from django.urls import path, include

from rest_framework.authtoken.views import obtain_auth_token

urlpatterns = [
    path('api-auth/', include('rest_framework.urls')),
    path('admin/', admin.site.urls),
    path('api/', include('core.urls')),
    path('api/token/', obtain_auth_token, name='obtain-token')
]
